Prevent error infinite loop when generating a UUID
Currently, we run `crypto.randomBytes(31).toString('hex')`, and if such
function throws any error, we wait for a millisecond and retry in order
to cope with low entropy situations.

The issue is that if the function above throws an error unrelated to
entropy, like:

	secure random number generation not supported by this browser use chrome, FireFox or Internet Explorer 11

which happens when running this function on IE 10, an infinite loop
occurs.